wireframes for unit 2
Here are two different thumbnail sketches of possible layouts for my
project Both are examples of one page sites as I feel it will be the
most user friendly format for a portfolio site as it presents the user
with all of the relevant content with the least amount of effort. I
also want to incorporate interactive features to allow the user to feel
engaged with the site and therefor the work in the portfolio.